+++ Raytracer/DielectricMaterial.swift
@@ -10,29 +10,29 @@ import Foundation
 import simd
 
 public class DielectricMaterial: Material {
-	public var indexOfRefraction: Float
-    public var attenuation: float3
-    public var fuzziness: Float
+	public var indexOfRefraction: Double
+    public var attenuation: double3
+    public var fuzziness: Double
     
-	init(refractionIndex: Float) {
+	init(refractionIndex: Double) {
 		indexOfRefraction = refractionIndex
-        attenuation = float3(1.0, 1.0, 1.0)
+        attenuation = double3(1.0, 1.0, 1.0)
         fuzziness = 0
 	}
     
-    init(refractionIndex: Float, color: float3, fuzziness f: Float) {
+    init(refractionIndex: Double, color: double3, fuzziness f: Double) {
         indexOfRefraction = refractionIndex
         attenuation = color
         fuzziness = f
     }
 	
-	public override func scatter(rayIn: Ray, rec: HitRecord) -> (attenuation: float3, scattered: Ray)? {
-		var outwardNormal: float3
+	public override func scatter(rayIn: Ray, rec: HitRecord) -> (attenuation: double3, scattered: Ray)? {
+		var outwardNormal: double3
 		let reflected = reflect(rayIn.direction, n: rec.normal)
-		var niOverNt: Float
-		var refracted: float3
-		var reflectionProbability: Float
-		var cosine: Float
+		var niOverNt: Double
+		var refracted: double3
+		var reflectionProbability: Double
+		var cosine: Double
 		
 		if dot(rayIn.direction, rec.normal) > 0 {
 			outwardNormal = -rec.normal
@@ -47,7 +47,7 @@ public class DielectricMaterial: Material {
 		refracted = refract(rayIn.direction, n: outwardNormal, eta: niOverNt)
 		
 		var scattered: Ray
-		let zero = float3()
+		let zero = double3()
 		
 		if refracted.x != zero.x && refracted.y != zero.y && refracted.z != zero.z {
 //			scattered = Ray(origin: rec.point, direction: refracted)
@@ -57,7 +57,7 @@ public class DielectricMaterial: Material {
 			reflectionProbability = 1.0
 		}
 		
-		if Float(drand48()) < reflectionProbability {
+		if drand48() < reflectionProbability {
 			scattered = Ray(origin: rec.point, direction: reflected + fuzziness * randomInUnitSphere(), time: rayIn.time)
 		} else {
 			scattered = Ray(origin: rec.point, direction: refracted + fuzziness * randomInUnitSphere(), time: rayIn.time)
